Reorganized Simulator and Tile code Split into multiple files for cleanliness
