+++ src/main.ts
@@ -1,7 +1,7 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import { AllLanguageData } from './constants';
+import { AllLanguageData } from './constants/language-data';
 
 import { createI18n } from 'vue-i18n'
